+++ packages/generator/CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.4
 - Use exit code 0 when there is no packages to publish.
+- Use annotated tags and use `--follow-tags` on push.
 
 ## 0.1.3
 - Init debug functions